+++ CHANGELOG.md
@@ -4,8 +4,9 @@ All notable changes to this package are documented in this file. This project ad
 
 ## Unreleased
 
-- #50, Fix wrong normalized path with descendant segment
 - #43, Fix spaces not allowed in relative query and singular query segments
+- #50, Fix wrong normalized path with descendant segment
+- #30, Fix escape characters not handled properly in normalized path
 
 ## 0.3.0.1